format, revert some changes
