+++ examples/current_user_recently_played.rs
@@ -1,4 +1,4 @@
-use rspotify::SpotifyBuilder;
+use rspotify::client::SpotifyBuilder;
 use rspotify::oauth2::{CredentialsBuilder, OAuthBuilder};
 use rspotify::scopes;
