+++ examples/ureq/search.rs
@@ -1,6 +1,7 @@
-use rspotify::{SpotifyBuilder, scopes};
+use rspotify::client::SpotifyBuilder;
 use rspotify::model::{Country, Market, SearchType};
 use rspotify::oauth2::{CredentialsBuilder, OAuthBuilder};
+use rspotify::scopes;
 
 use std::collections::HashSet;
