+++ examples/webapp/src/main.rs
@@ -14,9 +14,9 @@ use rocket::response::Redirect;
 use rocket_contrib::json;
 use rocket_contrib::json::JsonValue;
 use rocket_contrib::templates::Template;
-use rspotify::{ClientError, SpotifyBuilder};
+use rspotify::client::{ClientError, SpotifyBuilder};
 use rspotify::oauth2::{CredentialsBuilder, OAuthBuilder, TokenBuilder};
-use rspotify::scope;
+use rspotify::scopes;
 
 use std::fs;
 use std::{
